feat(DSTEW-1595): - Now uses MockServerIntegrationTest in ClaimAmendmentRepricingIntegrationTest - Moved changes from the ClaimControllerIntegrationTest
